+++ src/lib/waku_relay/constants.ts
@@ -4,7 +4,10 @@ export const minute = 60 * second;
 /**
  * RelayCodec is the libp2p identifier for the waku relay protocol
  */
-export const RelayCodec = '/vac/waku/relay/2.0.0-beta2';
+export const RelayCodecs = [
+  '/vac/waku/relay/2.0.0-beta2',
+  '/vac/waku/relay/2.0.0',
+];
 
 /**
  * DefaultPubsubTopic is the default gossipsub topic to use for Waku.